requirements.txt
Django==5.2.3
fhir.resources>=8.2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djangorestframework-camel-case>=1.4.2
gunicorn==20.1.0
whitenoise==6.5.0
django-cors-headers==4.7.0
psycopg2-binary>=2.9.7
drf-yasg==1.21.5
django-environ
dj-database-url
boto3>=1.34.0
django-storages==1.14.2
qrcode[pil]==8.2
pyotp==2.9.0
python-dotenv>=1.0.0
Pillow==11.3.0
django-ses>=3.5.0
bleach